requirements.txt
# Core
beautifulsoup4>=4.12.0
lxml>=4.9.0
edinet-xbrl>=0.2.0
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
streamlit>=1.32.0

# Phase 2: 決算説明資料 (PDF/PPTX)
pdfminer.six>=20240706
python-pptx>=1.0.0

# Phase 2 optional: Google Drive 同期
# google-api-python-client>=2.120.0
# google-auth>=2.29.0

# Phase 3 optional: xlsx / 画像レンダリング (実装時に有効化)
# openpyxl>=3.1.0
# pdf2image>=1.17.0
# pymupdf>=1.24.0
